requirements.txt
streamlit
rdflib
google-generativeai
python-dotenv
pyvis
